# docs/swagger.yaml
definitions:
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.Post:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      dislikes:
        type: integer
      forum_id:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_flagged:
        type: boolean
      likes:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      password:
        type: string
      uid:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /public/forgot-password:
    post:
      consumes:
      - application/json
      description: Envía un enlace de recuperación de contraseña al correo electrónico
        proporcionado.
      parameters:
      - description: Correo electrónico del usuario
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Enlace de recuperación enviado correctamente
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Solicitud inválida: el email es obligatorio'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno al enviar el enlace de recuperación
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recuperación de contraseña
      tags:
      - Auth
  /public/posts:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de todas las publicaciones ordenadas por fecha
        de creación.
      produces:
      - application/json
      responses:
        "200":
          description: Lista de publicaciones
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener todas las publicaciones
      tags:
      - Post
    post:
      consumes:
      - multipart/form-data
      description: Permite crear una nueva publicación con un título, contenido y
        una imagen opcional. La imagen se sube a Cloudinary y se guarda la URL en
        la publicación.
      parameters:
      - description: Título de la publicación
        in: formData
        name: title
        required: true
        type: string
      - description: Contenido de la publicación
        in: formData
        name: content
        required: true
        type: string
      - description: Imagen para la publicación
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Publicación creada exitosamente
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Solicitud inválida, título o contenido faltante
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno al crear la publicación
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear una nueva publicación
      tags:
      - Post
  /public/register:
    post:
      consumes:
      - application/json
      description: Permite registrar un nuevo usuario con su correo y contraseña
      parameters:
      - description: Datos del usuario a registrar
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario creado exitosamente
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Solicitud incorrecta: los datos no son válidos'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno al registrar el usuario
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar un nuevo usuario
      tags:
      - Auth
  /public/users:
    get:
      consumes:
      - application/json
      description: Recupera un usuario de la base de datos utilizando su ID.
      parameters:
      - description: ID del usuario a recuperar
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuario encontrado
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Parámetro 'id' faltante
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuario no encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener un usuario por ID
      tags:
      - User
swagger: "2.0"